updating cache_redis.php to fail silently
The current method used by the cache driver to check if a driver is_supported() only checks if the extension is loaded. In this version, I have moved the redis loading logic out of the constructor and into it's own load_driver() method. This allows the is_supported() method to also check if redis can be loaded successfully and if NOT, return false, which will tell the cache driver to failover silently to the backup method. For backwards compatibility, the load_driver() function will still return true, and allow things to fail if the config parameter 'ignore_on_fail' is left as false. This would be in cases where you want things to fail if the redis service can't be reached. Hopefully this doesn't seem to hacky. It works well for me.